+++ temportal-graphics/src/buffer/builder.rs
@@ -32,8 +32,8 @@ impl Builder {
 		self
 	}
 
-	pub fn with_size(mut self, size: u64) -> Self {
-		self.size = size;
+	pub fn with_size(mut self, size: usize) -> Self {
+		self.size = size as u64;
 		self
 	}
